+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/TextField.ts
@@ -10,6 +10,7 @@ import {
   AlloyEvents,
   AlloyTriggers,
   Behaviour,
+  Disabling,
   FormField as AlloyFormField,
   Input as AlloyInput,
   Invalidating,
@@ -18,21 +19,22 @@ import {
   Representing,
   SketchSpec,
   Tabstopping,
-  SystemEvents,
+  SystemEvents
 } from '@ephox/alloy';
+import { Types } from '@ephox/bridge';
 import { Arr, Future, Option, Result } from '@ephox/katamari';
 import { Traverse } from '@ephox/sugar';
 import { renderFormFieldWith, renderLabel } from 'tinymce/themes/silver/ui/alien/FieldLabeller';
 
 import { UiFactoryBackstageProviders } from '../../backstage/Backstage';
 import { formChangeEvent, formSubmitEvent } from '../general/FormEvents';
-import { Types } from '@ephox/bridge';
 import { Omit } from '../Omit';
 
 const renderTextField = function (spec: TextField, providersBackstage: UiFactoryBackstageProviders) {
   const pLabel = spec.label.map((label) => renderLabel(label, providersBackstage));
 
   const baseInputBehaviours = [
+    Disabling.config({ disabled: spec.disabled }),
     Keying.config({
       mode: 'execution',
       useEnter: spec.multiline !== true,
@@ -89,7 +91,19 @@ const renderTextField = function (spec: TextField, providersBackstage: UiFactory
 
   const extraClasses = spec.flex ? ['tox-form__group--stretched'] : [];
 
-  return renderFormFieldWith(pLabel, pField, extraClasses);
+  const extraBehaviours = [
+    Disabling.config({
+      disabled: spec.disabled,
+      onDisabled: (comp) => {
+        AlloyFormField.getField(comp).each(Disabling.disable);
+      },
+      onEnabled: (comp) => {
+        AlloyFormField.getField(comp).each(Disabling.enable);
+      }
+    })
+  ];
+
+  return renderFormFieldWith(pLabel, pField, extraClasses, extraBehaviours);
 };
 
 export type Validator = (v: string) => true | string;
@@ -101,6 +115,7 @@ export interface TextField {
   flex: boolean;
   label: Option<string>;
   placeholder: Option<string>;
+  disabled: boolean;
   validation: Option<{
     validator: Validator;
     validateOnLoad?: boolean
@@ -118,6 +133,7 @@ const renderInput = (spec: InputSpec, providersBackstage: UiFactoryBackstageProv
     label: spec.label,
     placeholder: spec.placeholder,
     flex: false,
+    disabled: spec.disabled,
     classname: 'tox-textfield',
     validation: Option.none()
   }, providersBackstage);
@@ -130,6 +146,7 @@ const renderTextarea = (spec: TextAreaSpec, providersBackstage: UiFactoryBacksta
     label: spec.label,
     placeholder: spec.placeholder,
     flex: true,
+    disabled: spec.disabled,
     classname: 'tox-textarea',
     validation: Option.none(),
   }, providersBackstage);