+++ modules/tinymce/src/themes/silver/test/ts/phantom/components/sizeinput/SizeInputTest.ts
@@ -16,7 +16,8 @@ UnitTest.asynctest('SizeInput component Test', (success, failure) => {
         renderSizeInput({
           name: 'dimensions',
           label: Option.some('size'),
-          constrain: true
+          constrain: true,
+          disabled: false
         }, TestProviders)
       );
     },